requirements.txt
boto3-stubs[dynamodb]
pytest==6.1.2
pytest-mock==3.3.1